El azzaoui Marwa & Tazi Ayoub/Projet/backend/requirements-correct.txt
Django==6.0.3
djangorestframework==3.15.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
mysql-connector-python==8.2.0
drf-spectacular==0.27.2
